// client/src/models/enums.ts
enum AviGrade {
    aviStart = 'AviStart',
    M3 = 'M3',
    E3 = 'E3',
    M4 = 'M4',
    E4 = 'E4',
    M5 = 'M5',
    E5 = 'E5',
    M6 = 'M6',
    E6 = 'E6',
    M7 = 'M7',
    E7 = 'E7',
    aviPlus = 'AviPlus',
}

enum FeedbackReviewStatus {
    pending = 'PENDING',
    reviewed = 'REVIEWED',
    processed = 'PROCESSED',
}

enum Severity {
    success = 'success',
    error = 'error',
    warning = 'warning',
    info = 'info',
}


export { AviGrade, FeedbackReviewStatus, Severity };